# backend/features/room/handler.go
package room

import (
	"encoding/json"
	"net/http"

	"github.com/Moeed-ul-Hassan/chatapp/core/authz"
	"github.com/Moeed-ul-Hassan/chatapp/core/middleware"
	"github.com/Moeed-ul-Hassan/chatapp/core/utils"
	"github.com/go-chi/chi/v5"
)

// respond writes a JSON response
func respond(w http.ResponseWriter, status int, data any) {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(status)
	json.NewEncoder(w).Encode(data)
}

// respondError writes a JSON error message
func respondError(w http.ResponseWriter, status int, msg string) {
	respond(w, status, map[string]string{"error": msg})
}

// RegisterRoutes attaches room REST APIs to the router
func RegisterRoutes(r chi.Router) {
	r.Group(func(r chi.Router) {
		r.Use(middleware.AuthRequired)
		r.Post("/", CreateRoomHandler)
		r.Get("/{roomID}", SearchRoomHandler)
		r.Post("/join", JoinRoomHandler)
	})
}

// CreateRoomHandler handles POST /api/rooms
// @Summary Create private room
// @Tags rooms
// @Security ApiKeyAuth
// @Accept json
// @Produce json
// @Param body body map[string]string true "name, passkey, aes_key"
// @Success 201 {object} Room
// @Failure 400 {object} map[string]string
// @Failure 401 {object} map[string]string
// @Failure 500 {object} map[string]string
// @Router /api/rooms/ [post]
func CreateRoomHandler(w http.ResponseWriter, r *http.Request) {
	claims := middleware.GetUserClaims(r)
	if claims == nil {
		respondError(w, http.StatusUnauthorized, "unauthorized")
		return
	}
	requesterUserID, ok := authz.ResolveLocalUserID(r)
	if !ok {
		respondError(w, http.StatusUnauthorized, "unable to resolve current user")
		return
	}

	var request struct {
		Name    string `json:"name"`
		Passkey string `json:"passkey"`
		AESKey  string `json:"aes_key"` // The E2EE master key generated by the client
	}

	if err := json.NewDecoder(r.Body).Decode(&request); err != nil {
		respondError(w, http.StatusBadRequest, "invalid request body")
		return
	}

	if request.Name == "" || request.Passkey == "" || request.AESKey == "" {
		respondError(w, http.StatusBadRequest, "name, passkey, and aes_key are required")
		return
	}

	hashedPasskey, err := utils.HashPassword(request.Passkey)
	if err != nil {
		respondError(w, http.StatusInternalServerError, "failed to process passkey")
		return
	}

	newRoom := &Room{
		Name:    request.Name,
		Passkey: hashedPasskey,
		AESKey:  request.AESKey,
		OwnerID: requesterUserID,
	}

	created, err := CreateRoom(r.Context(), newRoom)
	if err != nil {
		respondError(w, http.StatusInternalServerError, "failed to create room")
		return
	}

	respond(w, http.StatusCreated, created)
}

// SearchRoomHandler handles GET /api/rooms/{roomID}
// @Summary Get room metadata
// @Tags rooms
// @Security ApiKeyAuth
// @Param roomID path string true "Room ID"
// @Success 200 {object} map[string]interface{}
// @Failure 400 {object} map[string]string
// @Failure 404 {object} map[string]string
// @Router /api/rooms/{roomID} [get]
func SearchRoomHandler(w http.ResponseWriter, r *http.Request) {
	roomID := chi.URLParam(r, "roomID")
	if roomID == "" {
		respondError(w, http.StatusBadRequest, "room ID is required")
		return
	}

	room, err := GetRoomByRoomID(r.Context(), roomID)
	if err != nil {
		respondError(w, http.StatusNotFound, "room not found")
		return
	}

	// We only return public metadata (ID, Name, InviteOnly)
	// We definitely do NOT return the hashed passkey or the AES Key yet.
	respond(w, http.StatusOK, map[string]any{
		"id":          room.ID,
		"room_id":     room.RoomID,
		"name":        room.Name,
		"invite_only": room.InviteOnly,
		"require_2fa": room.Require2FA,
	})
}

// JoinRoomHandler handles POST /api/rooms/join
// Verifies passkey and returns the encrypted AES Key if successful.
// @Summary Join private room
// @Tags rooms
// @Security ApiKeyAuth
// @Accept json
// @Produce json
// @Param body body map[string]string true "room_id, passkey"
// @Success 200 {object} map[string]interface{}
// @Failure 400 {object} map[string]string
// @Failure 401 {object} map[string]string
// @Failure 404 {object} map[string]string
// @Router /api/rooms/join [post]
func JoinRoomHandler(w http.ResponseWriter, r *http.Request) {
	var request struct {
		RoomID  string `json:"room_id"`
		Passkey string `json:"passkey"`
	}

	if err := json.NewDecoder(r.Body).Decode(&request); err != nil {
		respondError(w, http.StatusBadRequest, "invalid request body")
		return
	}

	room, err := GetRoomByRoomID(r.Context(), request.RoomID)
	if err != nil {
		respondError(w, http.StatusNotFound, "room not found")
		return
	}

	isValid, err := utils.VerifyPassword(request.Passkey, room.Passkey)
	if err != nil || !isValid {
		respondError(w, http.StatusUnauthorized, "incorrect passkey")
		return
	}

	// In a full implementation, we might issue a room-specific session token.
	// For now, we return the AESKey which the client uses to decrypt messages.
	respond(w, http.StatusOK, map[string]any{
		"message": "access granted",
		"aes_key": room.AESKey,
	})
}